+++ src/spec2nexus/dev_spec.py
@@ -43,10 +43,10 @@ def developer_test(spec_file_name = None):
         #spec_file_name = os.path.join(spec_dir, '130123B_2.spc')
         spec_file_name = os.path.join(spec_dir, 'user6idd.dat')
         os.chdir(spec_dir)
-    print '-'*70
+        print '-'*70
     # now open the file and read it
     test = spec.SpecDataFile(spec_file_name)
-    scan = test.scans[2]
+    scan = test.getScan(1)
     scan.interpret()
     #print scan.UXML_root
     #print prettify(scan.UXML_root)
@@ -81,9 +81,6 @@ def developer_test(spec_file_name = None):
 
 
 if __name__ == "__main__":
-    fname = 'test_3.spec'
-    spec_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), 'data', 'uxml', ))
-    os.environ['SPEC2NEXUS_PLUGIN_PATH'] = spec_dir
-    full_name = os.path.join(spec_dir, fname)
     full_name = None
+    full_name = 'data/xpcs_plugin_sample.spec'
     developer_test(full_name)