+++ src/spec2nexus/writer.py
@@ -24,9 +24,9 @@ import h5py
 
 # see: http://download.nexusformat.org/doc/html/classes/base_classes/index.html
 #CONTAINER_CLASS = 'NXlog'          # information that is recorded against time
-#CONTAINER_CLASS = 'NXnote'         # store additional information in a NeXus file
+CONTAINER_CLASS = 'NXnote'         # any additional freeform information not covered by the other base classes
 #CONTAINER_CLASS = 'NXparameters'   # Container for parameters, usually used in processing or analysis
-CONTAINER_CLASS = 'NXcollection'    # Use NXcollection to gather together any set of terms
+#CONTAINER_CLASS = 'NXcollection'    # Use NXcollection to gather together any set of terms
         
 
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
@@ -55,6 +55,7 @@ class Writer(object):
         
             hdf5_file: NXroot
                 @default="S1"
+                definition="NXspecdata"
                 # attributes
                 S1:NXentry
                     @default="data"
@@ -77,12 +78,15 @@ class Writer(object):
         pick_first_entry = True
         for key in scan_list:
             nxentry = eznx.makeGroup(root, 'S'+str(key), 'NXentry')
+            eznx.makeDataset(nxentry, 
+                             'definition', 
+                             'NXspecdata', 
+                             description='NeXus application definition (status pending)')
             self.save_scan(nxentry, self.spec.getScan(key))
             if pick_first_entry:
                 pick_first_entry = False
                 eznx.addAttributes(root, default='S'+str(key))
             if 'data' not in nxentry:
-                # resolve issue #45 here
                 # NXentry MUST have a NXdata group with data for default plot
                 nxdata = eznx.makeGroup(nxentry, 
                                         'data', 
@@ -113,7 +117,9 @@ class Writer(object):
             SPEC_date = utils.iso8601(header0.date),
             SPEC_comments = '\n'.join(header0.comments),
             SPEC_num_headers = len(self.spec.headers),
-            h5py_version = h5py.__version__
+            h5py_version = h5py.__version__,
+            HDF5_Version = h5py.version.hdf5_version,
+            numpy_version = h5py.version.numpy.version.full_version,
             )
         try:
             c = header0.comments[0]
@@ -191,12 +197,9 @@ class Writer(object):
         return signal, axis
     
     def mca_spectra(self, nxdata, scan, primary_axis_label):
-        '''*internal*: parse for optional MCA spectra'''
-        if '_mca_' in scan.data:        # check for it
-            axes = primary_axis_label + ':' + '_mca_channel_'
-            channels = np.arange(1, len(scan.data['_mca_'][0])+1, dtype=int)
-            data = scan.data['_mca_']
-            self.write_ds(nxdata, '_mca_', data, axes=axes)
+        '''*internal*: parse for optional 2-D MCA spectra'''
+        if '_mca_' in scan.data:
+            # calibration for all spectra
             a, b, c = 0, 0, 0
             if hasattr(scan, 'MCA'):
                 mca = scan.MCA
@@ -206,8 +209,21 @@ class Writer(object):
                     c = mca['CALIB'].get('c', 0)
             if a == b and b == c and a == 0:
                 a, b, c = 1, 0, 0
-            _mca_x_ = a + channels * (b + channels * c)
-            self.write_ds(nxdata, '_mca_channel_', channels, units = 'channel')
+
+            # save each spectrum
+            for key, spectrum in sorted(scan.data['_mca_'].items()):
+                ds_name = '_' + key + '_'
+                axes = primary_axis_label + ':' + ds_name + 'channel_'
+                self.write_ds(nxdata, ds_name, spectrum, axes=axes, units = 'counts')
+                
+                # save calibrated channel data for each spectrum, in case spectra are different lengths
+                # _mca_     _mca_channel_
+                # _mca1_    _mca1_channel_
+                # _mca2_    _mca2_channel_
+                # ...
+                channels = np.arange(1, len(spectrum[0])+1, dtype=int)
+                _mca_x_ = a + channels * (b + channels * c)
+                self.write_ds(nxdata, ds_name + 'channel_', channels, units = 'channel')
     
     def mesh(self, nxdata, scan):
         '''*internal*: data parser for 2-D mesh and hklmesh'''
@@ -258,14 +274,17 @@ class Writer(object):
             signal = utils.clean_name(scan.column_last)
             axes = ':'.join([label1, label2])
 
-        if '_mca_' in scan.data:    # 3-D array
-            num_channels = len(scan.data['_mca_'][0])
-            data_shape.append(num_channels)
-            mca = np.array(scan.data['_mca_'])
-            data = utils.reshape_data(mca, data_shape)
-            channels = range(1, num_channels+1)
-            self.write_ds(nxdata, '_mca_', data, axes=axes+':'+'_mca_channel_')
-            self.write_ds(nxdata, '_mca_channel_', channels, units='channel')
+        if '_mca_' in scan.data:    # 3-D array(s)
+            # save each spectrum
+            for key, spectrum in sorted(scan.data['_mca_'].items()):
+                num_channels = len(spectrum[0])
+                data_shape.append(num_channels)
+                mca = np.array(spectrum)
+                data = utils.reshape_data(mca, data_shape)
+                channels = range(1, num_channels+1)
+                ds_name = '_' + key + '_'
+                self.write_ds(nxdata, ds_name, data, axes=axes+':'+ds_name+'channel_', units='counts')
+                self.write_ds(nxdata, ds_name+'channel_', channels, units='channel')
 
         return signal, axes